Refactoring ImageLayer
The new ImageLayer now uses Qt's SceneGraph to render the image. With
this, we get an improve on memory and CPU usage.

As consequence, some features were removed, as they are not needed
anymore (like the TiledDraw draw type).

But still needing work. They will be added to the Bacon2D issue tracker
[1] for a future major update.

Known issues: resizing is letting some dirty.

[1] https://github.com/Bacon2D/Bacon2D/issues/milestones

Signed-off-by: Rodrigo Goncalves de Oliveira <rodrigo.goncalves@openbossa.org>