Make running examples more convenient.
This patch adds .pro files for every example, using a common template
found under examples/shared to reduce code duplication. The QML,
images and sounds are added to a dynamically generated resource file,
which allows for shadow builds. The paths to these assets don't have
to be changed in the QML files, because relative URLs in QML assume
the scheme in use by the component that they're in. That is, if the
main.qml file for any given example is run by qmlscene, the URL will
use the file scheme, and if it is compiled into the application via a
QRC file, it will use the qrc scheme.

As the .pro-based example projects (as opposed to the existing
.qmlproject-based ones) use QQmlApplicationEngine, the main QML files
must now use Window as their root item [1]. This doesn't affect usage
with qmlscene.

Some bits of code had to be changed:

- References to variables that were previously in the "root" scope -
  the "game" item - were broken when "game" was made a child of the
  window.
- Use of game.width/game.height in Component.onCompleted of nested
  items were replaced with window.width/window.height, as we now fill
  "game" to the window, so its size is no longer known at that time.
- The viewport_parallax example needed a hard-coded size for "game";
  filling it to the window wasn't sufficient to get the ImageLayers
  to show up for some reason, which seems like a bug.

[1] http://doc.qt.io/qt-5/qqmlapplicationengine.html#details